pyproject.toml
[tool.poetry]
name = "plasticityhub"
version = "0.0.1"
description = "PlasticityHub is a Django-based database designed to manage and organize data for the Strauss Neuroplasticity Brain Bank. This platform provides a centralized, user-friendly interface for storing, querying, and tracking neuroimaging datasets, participant records, and metadata. Built with scalability and data integrity in mind, PlasticityHub streamlines the workflow for researchers by integrating customizable data management tools and ensuring efficient access to key research assets, supporting longitudinal studies on brain plasticity and neuroplasticity."
authors = ["GalKepler <galkepler@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "5.1"
python-environ = "^0.4.54"
whitenoise = "6.7.0"
django-crispy-forms = "2.3"
crispy-bootstrap5 = "2024.2"
python-slugify = "8.0.4"
pillow = "10.4.0"
argon2-cffi = "23.1.0"
redis = "5.0.8"
hiredis = "3.0.0"
django-environ = "0.11.2"
django-model-utils = "4.5.1"
django-allauth = {version = "64.1.0", extras = ["mfa"]}
django-redis = "5.4.0"
werkzeug = {version = "3.0.4", extras = ["watchdog"]}
ipdb = "0.13.13"
psycopg = {version = "3.2.1", extras = ["c"]}
mypy = "1.11.2"
django-stubs = {version = "5.0.4", extras = ["compatible-mypy"]}
pytest = "8.3.2"
pytest-sugar = "1.0.0"
sphinx = "8.0.2"
sphinx-autobuild = "2024.4.16"
ruff = "0.6.2"
coverage = "7.6.1"
djlint = "1.34.1"
pre-commit = "3.8.0"
factory-boy = "3.3.1"
django-debug-toolbar = "4.4.6"
django-extensions = "3.2.3"
django-coverage-plugin = "3.1.0"
pytest-django = "4.8.0"
gunicorn = "23.0.0"
django-anymail = {version = "11.1", extras = ["mailgun"]}
jupyter = "^1.0.0"
notebook = "^7.2.2"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"
isort = "^5.13.2"
black = "^24.8.0"
gspread = "^6.1.2"
gspread-dataframe = "^4.0.0"
tox = "^4.18.0"
sphinx-rtd-theme = {git = "https://github.com/readthedocs/sphinx_rtd_theme.git"}
djangoviz = "^0.1.1"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = [
    "tests.py",
    "test_*.py",
]

# ==== Coverage ====
[tool.coverage.run]
include = ["plasticityhub/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
target-version = "py312"
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    "*/migrations/*.py",
    "staticfiles/*",
]

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "C90",
  "I",
  "N",
  "UP",
  "YTT",
  # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
  "ASYNC",
  "S",
  "BLE",
  "FBT",
  "B",
  "A",
  "COM",
  "C4",
  "DTZ",
  "T10",
  "DJ",
  "EM",
  "EXE",
  "FA",
  'ISC',
  "ICN",
  "G",
  'INP',
  'PIE',
  "T20",
  'PYI',
  'PT',
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "SIM",
  "TID",
  "TCH",
  "INT",
  # "ARG", # Unused function argument
  "PTH",
  "ERA",
  "PD",
  "PGH",
  "PL",
  "TRY",
  "FLY",
  # "NPY",
  # "AIR",
  "PERF",
  # "FURB",
  # "LOG",
  "RUF",
]
ignore = [
  "S101", # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "SIM102", # sometimes it's better to nest
  "UP038", # Checks for uses of isinstance/issubclass that take a tuple
          # of types for comparison.
          # Deactivated because it can make the code slow:
          # https://github.com/astral-sh/ruff/issues/7871
]
# The fixes in extend-unsafe-fixes will require
# provide the `--unsafe-fixes` flag when fixing.
extend-unsafe-fixes = [
    "UP038",
]

[tool.ruff.lint.isort]
force-single-line = true

# black configuration
[tool.black]
line-length = 88
target-version = ['py310']

# isort configuration
[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3  # This setting allows multi-line imports with parentheses
include_trailing_comma = true  # Ensures trailing commas for better diff
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
